Move method into server.go
